# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Book Worm E-Bookstore API
  description: >
    REST API for the Book Worm e-commerce platform enabling customers to browse,
    select and purchase books. Covers authentication, catalogue browsing,
    shopping cart, wishlist, order management, payment processing and shipment tracking.
  version: 1.0.0
  contact:
    name: Book Worm Platform Team
servers:
  - url: https://api.bookworm.com/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Local development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  schemas:
...
    Role:
      type: string
      enum: [GUEST, MEMBER, ADMIN]
    BookFormat:
      type: string
      enum: [PAPERBACK, HARDCOVER, EBOOK]
    OrderStatus:
      type: string
      enum: [PENDING, CONFIRMED, CANCELLED, RETURNED, SHIPPED, DELIVERED]
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        giftPointBalance:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RegisterRequest:
      type: object
      required: [email, password, firstName, lastName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          description: Plain-text password — hashed server-side before storage
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT bearer token
        tokenType:
          type: string
          default: Bearer
        user:
          $ref: '#/components/schemas/UserResponse'
    UpdateUserRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string

    # ── Store ─────────────────────────────────────────────────────────────────
...
    BookSummary:
      type: object
      description: Lightweight book representation used in lists and recommendations
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        authorName:
          type: string
        coverImageUrl:
          type: string
          format: uri
        price:
          type: number
          format: double
        currencyCode:
          type: string
          default: INR
        format:
          $ref: '#/components/schemas/BookFormat'
        tags:
          type: array
          items:
            type: string
        tentativeDeliveryDate:
          type: string
          format: date
        averageRating:
          type: number
          format: double
...
    CartItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        book:
          $ref: '#/components/schemas/BookSummary'
        quantity:
          type: integer
          minimum: 1
        unitPrice:
          type: number
          format: double
    Cart:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AddToCartRequest:
      type: object
      required: [bookId, quantity]
      properties:
        bookId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
    UpdateCartItemRequest:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
          minimum: 1

    # ── Wishlist ──────────────────────────────────────────────────────────────
...
    Address:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        addressLine1:
          type: string
        addressLine2:
          type: string
        city:
          type: string
        pin:
          type: string
        state:
          type: string
        country:
          type: string
          default: India
        isSaved:
          type: boolean
    CreateAddressRequest:
      type: object
      required: [firstName, lastName, email, phone, addressLine1, city, pin, state, country]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        addressLine1:
          type: string
        addressLine2:
          type: string
        city:
          type: string
        pin:
          type: string
        state:
          type: string
        country:
          type: string
          default: India
        isSaved:
          type: boolean
          default: false

    # ── Order ─────────────────────────────────────────────────────────────────
    OrderItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        book:
          $ref: '#/components/schemas/BookSummary'
        quantity:
          type: integer
          minimum: 1
        unitPrice:
          type: number
          format: double
        tentativeDeliveryDate:
          type: string
          format: date
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        deliveryAddress:
          $ref: '#/components/schemas/Address'
        status:
          $ref: '#/components/schemas/OrderStatus'
        subtotal:
          type: number
          format: double
        tax:
          type: number
          format: double
        deliveryCharge:
          type: number
          format: double
        discount:
          type: number
          format: double
        totalAmount:
          type: number
          format: double
        couponCode:
          type: string
          nullable: true
        giftPointsRedeemed:
          type: integer
          minimum: 0
        paymentId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        cancellationDeadline:
          type: string
          format: date-time
    CreateOrderRequest:
      type: object
      required: [cartId, deliveryAddress]
      properties:
        cartId:
          type: string
          format: uuid
        deliveryAddress:
          $ref: '#/components/schemas/CreateAddressRequest'
        couponCode:
          type: string
        giftPointsToRedeem:
          type: integer
          minimum: 0

    # ── Payment ───────────────────────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Invalid request payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

# ── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate user and obtain JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /users/{userId}:
    get:
      tags: [Users]
      summary: Get user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Users]
      summary: Update user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /cart:
    get:
      tags: [Cart]
      summary: Get current user cart
      responses:
        '200':
          description: Cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags: [Cart]
      summary: Clear entire cart
      responses:
        '204':
          description: Cart cleared
        '401':
          $ref: '#/components/responses/Unauthorized'
  /cart/items:
    post:
      tags: [Cart]
      summary: Add a book to the cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddToCartRequest'
      responses:
        '201':
          description: Item added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /cart/items/{itemId}:
    put:
      tags: [Cart]
      summary: Update cart item quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCartItemRequest'
      responses:
        '200':
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Cart]
      summary: Remove a specific item from the cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Updated cart after removal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # Wishlist
...
  /orders:
    post:
      tags: [Orders]
      summary: Create a new order from cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      tags: [Orders]
      summary: List orders for current user
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Order list
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
  /orders/{orderId}:
    get:
      tags: [Orders]
      summary: Get order detail
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /orders/{orderId}/cancel:
    post:
      tags: [Orders]
      summary: Cancel an order (allowed within 48 hours of creation)
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...